apps/api/pyproject.toml
[project]
name = "knowledge-cloud-api"
version = "0.1.0"
description = "Private cloud knowledge base API"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlmodel>=0.0.24",
  "pydantic-settings>=2.10.1",
  "PyJWT>=2.9.0",
  "httpx>=0.28.1",
  "readability-lxml>=0.8.1",
  "markdownify>=0.14.1",
  "beautifulsoup4>=4.12.3",
  "cryptography>=43.0.1",
  "python-docx>=1.1.2",
  "python-pptx>=1.0.2",
  "pypdf>=5.4.0",
  "python-multipart>=0.0.20"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
